Restructure LineChart
