+++ projects/js-packages/charts/src/components/line-chart/line-chart.tsx
@@ -25,9 +25,7 @@ import { DefaultGlyph } from '../private/default-glyph';
 import { SingleChartContext, type SingleChartRef } from '../private/single-chart-context';
 import { withResponsive } from '../private/with-responsive';
 import { AccessibleTooltip, useKeyboardNavigation } from '../tooltip/private/accessible-tooltip';
-import LineChartAnnotation from './line-chart-annotation';
-import LineChartAnnotationsOverlay from './line-chart-annotations-overlay';
-import styles from './line-chart.module.scss';
+import { LineChartAnnotation, LineChartAnnotationsOverlay, styles } from './private';
 import type { BaseChartProps, DataPoint, DataPointDate, SeriesData, Optional } from '../../types';
 import type { ResponsiveConfig } from '../private/with-responsive';
 import type { TickFormatter } from '@visx/axis';
