+++ projects/js-packages/charts/src/components/line-chart/private/line-chart-annotation-label-popover.tsx
@@ -2,9 +2,9 @@ import { __ } from '@wordpress/i18n';
 import clsx from 'clsx';
 import Gridicon from 'gridicons';
 import { useEffect, useId, useRef, useState } from 'react';
-import { isSafari } from '../../utils';
+import { isSafari } from '../../../utils';
 import styles from './line-chart.module.scss';
-import type { ButtonWithPopover, PopoverElement, ToggleEvent } from '../../types';
+import type { ButtonWithPopover, PopoverElement, ToggleEvent } from '../../../types';
 import type { FC } from 'react';
 
 export const POPOVER_BUTTON_SIZE = 44;
