+++ projects/js-packages/charts/src/components/line-chart/private/line-chart.module.scss
